PNG encoder
Add a PNG data encoder with appropriate tests. Unlike PNG, JPEG and GIF
are lossy and are not able to be used without some hackery, modifying
the stdlib encoders for these filetypes.

A nice follow-up to this PR would be a stegoPNG encoder, which takes PNG
images and hides data in the LSBs of each pixel. May circle back and add
this in the future.