Improve Last Check-In for Sessions
As described in #197, there are issues with how LastCheckin is updated.

This adds a helper method to update LastCheckin to ensure it's done in
a consistent fashion, and calls it on all successful commands as well as
pings via the HTTP and DNS listeners.  Persistent connections (e.g.,
mtls) only update on commands, so will appear stale otherwise.

It no longer fills in LastCheckin for persistent connections when
someone runs the `sessions` command (or other grpc requests), so we
don't get overly-optimistic versions of the LastCheckin.